cambiados los datatypes de mi parte

+++ api/models/administradorBD__Model.js
@@ -5,11 +5,11 @@ const AdministradorBD = sequelize.define(
     'AdministradorBD',
     {
         id_Administador: {
-            type: DataTypes.SMALLINT.UNSIGNED,
+            type: DataTypes.TINYINT.UNSIGNED,
             primaryKey: true,
             autoIncrement: true
         },
-        id_Rol: DataTypes.MEDIUMINT.UNSIGNED,
+        id_Rol: DataTypes.TINYINT.UNSIGNED,
         nombre: {
             type: DataTypes.STRING(15),
             validate: {
