+++ api/models/productoCandy__Model.js
@@ -12,7 +12,7 @@ const ProductoCandy = sequelize.define(
         stock: DataTypes.MEDIUMINT.UNSIGNED,
         nombre: DataTypes.STRING(15),
         categoria: DataTypes.STRING(15),
-        precio: DataTypes.INTEGER.UNSIGNED,
+        precio: DataTypes.MEDIUMINT.UNSIGNED,
         DVH: {
             type: DataTypes.STRING(6),
             allowNull: false,
